Use size_T for memory block sizes

+++ StbImage/StbImage.h
@@ -6,5 +6,5 @@
 
 #define DLLEXPORT __declspec(dllexport)
 DLLEXPORT int GetImageInfo(char const* filename, int* width, int* height, int* numComponents);
-DLLEXPORT int ReadImageAsBCx(char const* filename, int flipVertically, int format, unsigned char* dest, int destSize);
-DLLEXPORT int ReadImageAsRGBA(char const* filename, int flipVertically, unsigned char* dest, int destSize);
+DLLEXPORT int ReadImageAsBCx(char const* filename, int flipVertically, int format, unsigned char* dest, size_t destSize);
+DLLEXPORT int ReadImageAsRGBA(char const* filename, int flipVertically, unsigned char* dest, size_t destSize);
